k8s: annotate k8s node after IP address allocation
In a k8s environment, the node annotation is fundamental for the
etcd-operator integration. Cilium uses the node annotation to create the
bpf tunnels between the nodes until it falls back to the kvstore.

As AnnotateNode runs inside a go routine, the IP address annotation
would race when executing ipam.Init().

Signed-off-by: André Martins <andre@cilium.io>